Disable downvoting for low reputation

+++ overboard/posts/views.py
@@ -116,18 +116,21 @@ def question_vote(request, question_id):
                 elif v.voter == user:
                     found_opposite_vote = True
                     found_vote = v
+            voter = UserExtended.objects.filter(user=user).first()
             if found_duplicate_vote or found_opposite_vote:
-                found_vote.delete()
-                vote_target_user = UserExtended.objects.filter(user=question.asked_by).first()
-                vote_target_user.reputation = vote_target_user.reputation - found_vote.value
-                vote_target_user.save()
+                if not (voter.reputation < 10 and value < 0):
+                    found_vote.delete()
+                    vote_target_user = UserExtended.objects.filter(user=question.asked_by).first()
+                    vote_target_user.reputation = vote_target_user.reputation - found_vote.value
+                    vote_target_user.save()
             if not found_duplicate_vote and user != question.asked_by:
                 current_date = datetime.datetime.now()
-                vote = Vote.objects.create(voter=user, vote_date=current_date, value=value, target=question)
-                vote.save()
-                vote_target_user = UserExtended.objects.filter(user=question.asked_by).first()
-                vote_target_user.reputation = vote_target_user.reputation + value
-                vote_target_user.save()
+                if not (voter.reputation < 10 and value < 0):
+                    vote = Vote.objects.create(voter=user, vote_date=current_date, value=value, target=question)
+                    vote.save()
+                    vote_target_user = UserExtended.objects.filter(user=question.asked_by).first()
+                    vote_target_user.reputation = vote_target_user.reputation + value
+                    vote_target_user.save()
         else:
             return HttpResponseRedirect('/404')
     return HttpResponseRedirect(reverse('posts:question_page', args=(question.id,)))
@@ -151,18 +154,21 @@ def answer_vote(request, question_id):
                 elif v.voter == user:
                     found_opposite_vote = True
                     found_vote = v
+            voter = UserExtended.objects.filter(user=user).first()
             if found_duplicate_vote or found_opposite_vote:
-                found_vote.delete()
-                vote_target_user = UserExtended.objects.filter(user=answer.published_by).first()
-                vote_target_user.reputation = vote_target_user.reputation - found_vote.value
-                vote_target_user.save()
+                if not (voter.reputation < 10 and value < 0):
+                    found_vote.delete()
+                    vote_target_user = UserExtended.objects.filter(user=answer.published_by).first()
+                    vote_target_user.reputation = vote_target_user.reputation - found_vote.value
+                    vote_target_user.save()
             if not found_duplicate_vote and user != answer.published_by:
-                current_date = datetime.datetime.now()
-                vote = Vote.objects.create(voter=user, vote_date=current_date, value=value, target=answer)
-                vote.save()
-                vote_target_user = UserExtended.objects.filter(user=answer.published_by).first()
-                vote_target_user.reputation = vote_target_user.reputation + value
-                vote_target_user.save()
+                if not (voter.reputation < 10 and value < 0):
+                    current_date = datetime.datetime.now()
+                    vote = Vote.objects.create(voter=user, vote_date=current_date, value=value, target=answer)
+                    vote.save()
+                    vote_target_user = UserExtended.objects.filter(user=answer.published_by).first()
+                    vote_target_user.reputation = vote_target_user.reputation + value
+                    vote_target_user.save()
     return HttpResponseRedirect(reverse('posts:question_page', args=(question.id,)))
